+++ settings/locales/en.json
@@ -151,5 +151,12 @@
 		"realtor": "portrait of the agent : {{realtor}}",
 		"agency": "view of the agency : {{agency}}",
 		"blog": "illustration of the blog : {{blog}}"
+	},
+	"maps": {
+		"error": {
+			"mapLoading": "Error while loading the map: {{error}}",
+			"addressNotFound": "Some addresses could not be located.",
+			"addressGeo": "Error during geocoding."
+		}
 	}
 }
