+++ settings/resources/luxe-jahia-demo.properties
@@ -29,6 +29,9 @@ luxe_estate.description = Description
 luxe_estate.description.ui.tooltip = Provide a description of the property and emphasize its key features and strengths.
 luxe_estate.price = Price
 luxe_estate.images = Image Gallery
+luxe_estate.address = Property address  
+luxe_estate.address.ui.tooltip = Full address of the property, including street, postal code, and city.  
+luxe_estate.state = State/Region
 luxe_estate.country = Country
 luxe_estate.country.ui.tooltip = The country where the property is located.
 luxe_estate.type = Type of Real Estate
