+++ settings/resources/luxe-jahia-demo_de.properties
@@ -29,6 +29,9 @@ luxe_estate.description = Beschreibung
 luxe_estate.description.ui.tooltip = Geben Sie eine Beschreibung der Immobilie an und heben Sie deren wesentliche Merkmale und St‰rken hervor.
 luxe_estate.price = Preis
 luxe_estate.images = Bildergalerie
+luxe_estate.address = Immobilienadresse  
+luxe_estate.address.ui.tooltip = Vollst‰ndige Adresse der Immobilie, einschlieﬂlich Straﬂe, Postleitzahl und Stadt.  
+luxe_estate.state = Region
 luxe_estate.country = Land
 luxe_estate.country.ui.tooltip = Das Land, in dem sich die Immobilie befindet.
 luxe_estate.type = Immobilientyp
