+++ settings/resources/luxe-jahia-demo_es.properties
@@ -29,6 +29,9 @@ luxe_estate.description = Descripción
 luxe_estate.description.ui.tooltip = Proporcione una descripción de la propiedad y resalte sus características y puntos fuertes.
 luxe_estate.price = Precio
 luxe_estate.images = Galería de imágenes
+luxe_estate.address = Dirección de la propiedad  
+luxe_estate.address.ui.tooltip = Dirección completa de la propiedad, incluyendo calle, código postal y ciudad.  
+luxe_estate.state = Región
 luxe_estate.country = País
 luxe_estate.country.ui.tooltip = El país donde se encuentra la propiedad.
 luxe_estate.type = Tipo de bien inmueble
