+++ settings/resources/luxe-jahia-demo_fr.properties
@@ -29,6 +29,9 @@ luxe_estate.description = Description
 luxe_estate.description.ui.tooltip = Fournissez une description de la propriété et mettez en avant ses caractéristiques clés.
 luxe_estate.price = Prix
 luxe_estate.images = Galerie d'images
+luxe_estate.address = Adresse du bien immobilier
+luxe_estate.address.ui.tooltip = Adresse complète de la propriété, y compris la rue, le code postal et la ville.
+luxe_estate.state = Région
 luxe_estate.country = Pays
 luxe_estate.country.ui.tooltip = Le pays où se situe la propriété.
 luxe_estate.type = Type de bien immobilier
