+++ src/commons/map/MapWithPin.client.module.css
@@ -16,12 +16,11 @@
 	bottom: 0;
 	left: 0;
 	line-height: 1.4;
-	background: rgba(255, 255, 255, 0.90); /* Bootstrap danger */
+	background: rgba(255, 255, 255, 0.8); /* Bootstrap danger */
 	color: rgba(220, 53, 69, 0.95); /* Bootstrap danger */
 	padding: 0 5px;
 	font-size: .75rem;
 	z-index: 400;
 	max-width: 70%;
-	box-shadow: 0 2px 8px rgba(0,0,0,0.12);
 	pointer-events: none; /* Allow map to receive mouse events */
 }
