+++ src/commons/map/MapWithPin.client.tsx
@@ -2,15 +2,9 @@ import React, { useEffect, useState } from "react";
 import { geocodeAddress } from "~/commons/map/geocodeAddress";
 import "leaflet/dist/leaflet.css";
 import classes from "~/commons/map/MapWithPin.client.module.css";
-import type { LeafletMapClientProps } from "~/commons/map/LeafletMap.client";
+import type { AddressItem, LeafletMapClientProps } from "~/commons/map/types";
 import clsx from "clsx";
-
-// const LeafletMapClient = React.lazy(() => import("~/commons/map/LeafletMap.client"));
-export type AddressItem = {
-	id: string;
-	label: string; // For popup (e.g. agency name or address)
-	address: string;
-};
+import { t } from "i18next";
 
 type Coordinates = {
 	lat: number;
@@ -28,16 +22,8 @@ const MapWithPinClient: React.FC<MapWithPinClientProps> = ({ addresses, classNam
 	const [coords, setCoords] = useState<Coordinates[]>([]);
 	const [LeafletMapClient, setLeafletMapClient] =
 		useState<React.ComponentType<LeafletMapClientProps> | null>(null);
-
-	// State to check if we are on client-side (for SSR compatibility)
-	// const [isClient, setIsClient] = useState(false);
 	// State for error message
 	const [error, setError] = useState<string | null>(null);
-
-	// useEffect(() => {
-	// 	// Ensure map is only rendered on client side
-	// 	setIsClient(true);
-	// }, []);
 
 	useEffect(() => {
 		// Import dynamically ONLY on client
@@ -45,7 +31,7 @@ const MapWithPinClient: React.FC<MapWithPinClientProps> = ({ addresses, classNam
 			.then((mod) => setLeafletMapClient(() => mod.default))
 			.catch((e) => {
 				console.error(e);
-				setError("Erreur lors du chargement de la carte.");
+				setError(t("maps.error.mapLoading", { error: e.message }));
 			});
 	}, []);
 
@@ -70,10 +56,10 @@ const MapWithPinClient: React.FC<MapWithPinClientProps> = ({ addresses, classNam
 				const filtered = results.filter((item): item is Coordinates => !!item);
 				setCoords(filtered);
 				if (filtered.length < addresses.length) {
-					setError("Certaines adresses n'ont pas pu être localisées.");
+					setError(t("maps.error.addressNotFound"));
 				}
 			})
-			.catch(() => setError("Erreur lors du géocodage."));
+			.catch(() => setError(t("maps.error.addressGeo")));
 	}, [addresses]);
 
 	// if (!isClient) return <div>Chargement…</div>; // Do not render on server
